docs : create rectangle zoom
